requirements.txt
requests==2.12.4
pytesseract==0.1.7
opencv_python==3.2.0.7
numpy==1.14.0
ipdb==0.10.2
mpld3==0.3
matplotlib==2.0.2
Image==1.5.17
Pillow==5.0.0
beautifulsoup4==4.6.0
